requirements.txt
langchain-openai
python-dotenv
langchain
langchain-core
scikit-learn
numpy
streamlit
pandas
plotly